% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{bind}
\alias{bind}
\title{Bind a function or tacit function to one or more of its arguments}
\usage{
bind(f, ...)
}
\arguments{
\item{f}{A function or tacit function}

\item{...}{values for arguments to be bound (can be more than one)}
}
\value{
A function or tacit function depending on the type of the argument f
}
\description{
Bind a function or tacit function to one or more of its arguments
}
\examples{
f<-function(a,b) a^b
g<-bind(f,2) # g(b) = f(2,b) = 2^b ... g is a function
g(3) # g(3) = 8
g<-bind(f,b=2) # g(a) = f(a,2) = a^2
g(3) # g(3) = 9
F<-I\%.\%f  # F(a,b) = f(a,b) ... F is a tacit function
G<-bind(F,b=2) # as before, but G is a tacit function
}
