% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{\%.\%}
\alias{\%.\%}
\title{Composition of functions and tacit functions}
\usage{
f \%.\% g
}
\arguments{
\item{f}{Function or tacit function}

\item{g}{Function or tacit function}
}
\value{
A tacit function
}
\description{
For functions or tacit functions f(x) and g(x), h<-f\%.\%g forms 
the composition h(x) where h(x) = f(g(x)) and h is a tacit 
function (contrast this with \%..\%).
}
\examples{
f <- sin \%.\% cos  # f(x) = sin(cos(x)) ... a tacit function
f <- I \%.\% exp      
g <- I \%.\% sin
h <- f * g \%.\% cos # h(x) = exp(x) * sin(cos(x)) ... a tacit function
h <- (f * g) \%.\% cos # h(x) = exp(cos(x)) * sin(cos(x)) ... a tacit function
}
