% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TACIT_function.R
\name{find_main_cell_types_with_subsets_and_markers}
\alias{find_main_cell_types_with_subsets_and_markers}
\title{Find blocks from PROTEINxCELLTYPE matrix}
\usage{
find_main_cell_types_with_subsets_and_markers(signature)
}
\arguments{
\item{signature}{a matrix where rows correspond to cell types and columns to markers.
The first column should be the name of the cell types.}
}
\value{
A list where each main cell type is the name of the list element.
Each element contains a character vector of additional subset markers not present in the main cell type.
}
\description{
This function analyzes a PROTEINxCELLTYPE matrix to extract main cell types and their associated markers.
It also identifies subset cell types that share all markers of a main cell type and have additional unique markers.
}
