% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TACIT_function.R
\name{heatmap_anb}
\alias{heatmap_anb}
\title{Create a Dot-Heatmap for Antibody Expressions}
\usage{
heatmap_anb(data_pos_neg, data_anb, group, name_anb)
}
\arguments{
\item{data_pos_neg}{a dataframe with binary (0/1) data indicating positive or negative expression.}

\item{data_anb}{a dataframe with quantitative data for antibody expressions.}

\item{group}{a vector indicating the group classification for each cell.}

\item{name_anb}{a vector of antibody names to order the heatmap.}
}
\value{
A ggplot object representing the dot-heatmap with points colored by mean value and sized by
percentage of cells expressing each antibody.
}
\description{
This function generates a heatmap to visualize the expression levels and frequency of antibodies.
It uses median values for expression levels and calculates the percentage of cells expressing each
antibody in different groups.
}
\examples{
# Example usage:
# Assuming 'data_pos_neg', 'data_anb', 'group', and 'name_anb' are already defined
heatmap <- createHeatmapForAntibodies(data_pos_neg, data_anb, group, name_anb)
print(heatmap)

}
