% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TACIT_function.R
\name{removeFarConnections}
\alias{removeFarConnections}
\title{Remove Far Connections in Triangulation}
\usage{
removeFarConnections(tri, cutoff)
}
\arguments{
\item{tri}{a triangulation object, typically from a Delaunay triangulation.}

\item{cutoff}{a numeric value between 0 and 1 indicating the quantile of edge distances to use as a cutoff.}
}
\value{
A modified triangulation object with edges longer than the cutoff distance removed.
}
\description{
This function removes edges from a triangulation object that are longer than a specified cutoff distance.
The cutoff is determined as a quantile of the distances between points in the triangulation.
}
\examples{
# Example usage:
# Assuming 'tri' is a triangulation object (e.g., from Delaunay triangulation)
cutoff <- 0.95
new_tri <- removeFarConnections(tri, cutoff)
print(new_tri)

}
