% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TACIT_function.R
\name{TACIT}
\alias{TACIT}
\title{Threshold-based Assignment of Cell Types from Multiplexed Imaging DaTa}
\usage{
TACIT(data_expression, r, p, Signature)
}
\arguments{
\item{data_expression}{a matrix or dataframe representing cell in the column and marker in the row.}

\item{r}{resolution value for microclusters.}

\item{p}{number of dimension used for microclusters.}

\item{Signature}{a matrix or dataframe representing signature of the cell type.}
}
\value{
A list with four elements:
\itemize{
\item{final_threshold}{The dataframe of threshold for each cell type relevant score.}
\item{Group_threshold_data}{The binary matrix using the threshold.}
\item{data_clean_final}{The dataframe of final annotation.}
\item{clusters}{The microclusters index for each cells.}
}
}
\description{
This function identifies cell type annotation using CELLxFEATURE matrix and Signature matrix in spatial omics data.
It uses segmented regression to find breakpoints in the median intensity values of cell type relevant score
across clusters.The function then assigns intensity groups based on these breakpoints and
calculates a threshold for each cell type relevant score. The KNN would be used in the feature subspace to gave final annotation.
}
\examples{
# Example usage:
# Assuming 'data_expression', and 'Signature' are already defined
annotation <- TACIT(data_expression, r, Signature, p)
print(annotations[3])

}
