% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TACIT_function.R
\name{TACIT_visualization}
\alias{TACIT_visualization}
\title{Visualize TACIT Data with UMAP Coordinates}
\usage{
TACIT_visualization(data_anb, coor_x, coor_y, cell_annotation)
}
\arguments{
\item{data_anb}{a dataframe or matrix containing marker intensity data or other relevant data for UMAP.}

\item{coor_x}{a numeric vector representing the x-coordinates in a spatial context.}

\item{coor_y}{a numeric vector representing the y-coordinates in a spatial context.}

\item{cell_annotation}{a vector or factor providing cell annotations.}
}
\value{
A dataframe that includes original x and y coordinates, UMAP coordinates, cell annotations,
and the original data used for UMAP. Column names are 'X', 'Y', 'UMAP1', 'UMAP2', 'TACIT',
followed by the original columns from \code{data_anb}.
}
\description{
This function applies UMAP (Uniform Manifold Approximation and Projection) to the provided data
and combines the resulting coordinates with existing spatial coordinates and cell annotations.
It's designed to prepare data for visualization in the context of TACIT analysis.
}
\examples{
# Example usage:
# Assuming 'data_anb', 'coor_x', 'coor_y', and 'cell_annotation' are already defined
visualization_data <- TACIT_visualization(data_anb, coor_x, coor_y, cell_annotation)
print(visualization_data)

}
