% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TACIT_function.R
\name{threshold_function}
\alias{threshold_function}
\title{Determine Thresholds for Cell Types Score and Markers}
\usage{
threshold_function(k, data_anb, clusters)
}
\arguments{
\item{k}{the index of the marker column in the \code{data_anb} dataframe.}

\item{data_anb}{a dataframe containing marker intensity data.}

\item{clusters}{a vector or factor of cluster assignments for each cell in \code{data_anb}.}
}
\value{
A list with two elements:
\itemize{
\item{Group_Threshold_new}{A binary matrix representing high (1) and low (0) expressions of markers.}
\item{threshold_point_final}{The calculated threshold value for marker binarization.}
}
}
\description{
This function identifies optimal threshold values for cell type score and markers in a CELLxPROTEIN and CELLxCELLTYPE matrix based on cluster analysis.
It uses segmented regression to find breakpoints in the median intensity values of markers across clusters.
The function then assigns intensity groups based on these breakpoints and calculates a threshold to
binary marker expressions into high (1) and low (0) intensity groups.
}
\examples{
# Example usage:
# Assuming 'data_anb' is your dataframe and 'clusters' is a vector of cluster assignments.
k <- 2 # Index of the marker column in 'data_anb'
result <- threshold_function(k, data_anb, clusters)
print(result)

}
