% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TACIT_function.R
\name{threshold_groups}
\alias{threshold_groups}
\title{Assign Groups Based on Thresholds}
\usage{
threshold_groups(input_vector, thresholds)
}
\arguments{
\item{input_vector}{a numeric vector to which group numbers will be assigned.}

\item{thresholds}{a numeric vector of threshold values used for grouping.
Each element in \code{input_vector} is assigned a group number based on these thresholds.}
}
\value{
A numeric vector of the same length as \code{input_vector},
with each element assigned a group number based on the specified thresholds.
}
\description{
This function assigns group numbers to elements of an input vector based on a set of threshold values.
Each element of the input vector is categorized into a group based on which thresholds it falls between.
The function supports multiple thresholds, allowing for a range of group assignments.
}
\examples{
# Example usage:
input_vector <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
thresholds <- c(3, 6, 9)
groups <- threshold_groups(input_vector, thresholds)
print(groups)

}
