% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TACIT_function.R
\docType{package}
\name{TACIT}
\alias{TACIT-package}
\alias{TACIT}
\title{TACIT package helper functions}
\description{
Package imports used by functions in this file. These tags make sure
the necessary namespaces are available when the package is built.
}
\author{
\strong{Maintainer}: Khoa Huynh \email{huynhk4@vcu.edu}

}
\keyword{internal}
